fix: Generate TS compiler error to stop unexpected behaviour.
The `arrayAction: 'overwrite'` and `arrayIndex` options are mutually exclusive, this patch increases verbosity but fixes this behaviour on compile time.

For JavaScript users... either use `// @ts-check`, or you're on your own. Either way this is more likely to affect advanced users, since they're the ones who will mess with those options, definitely not the ones who are new to SettingsGateway.